client context out of utils
