+++ src/viam/sdk/common/client_helper.cpp
@@ -2,7 +2,11 @@
 
 #include <cstdlib>
 
+#include <grpcpp/client_context.h>
+
 #include <boost/log/trivial.hpp>
+
+#include <viam/sdk/common/private/version_metadata.hpp>
 
 namespace viam {
 namespace sdk {
@@ -16,5 +20,37 @@ namespace client_helper_details {
 }
 
 }  // namespace client_helper_details
+
+ClientContext::ClientContext() : wrapped_context_(std::make_unique<grpc::ClientContext>()) {
+    set_client_ctx_authority_();
+    add_viam_client_version_();
+}
+
+ClientContext::~ClientContext() = default;
+
+ClientContext::operator const grpc::ClientContext*() const {
+    return wrapped_context_.get();
+}
+
+ClientContext::operator grpc::ClientContext*() {
+    return wrapped_context_.get();
+}
+
+void ClientContext::try_cancel() {
+    wrapped_context_->TryCancel();
+}
+
+void ClientContext::set_debug_key(const std::string& debug_key) {
+    wrapped_context_->AddMetadata("dtname", debug_key);
+}
+
+void ClientContext::set_client_ctx_authority_() {
+    wrapped_context_->set_authority("viam-placeholder");
+}
+
+void ClientContext::add_viam_client_version_() {
+    wrapped_context_->AddMetadata("viam_client", impl::k_version);
+}
+
 }  // namespace sdk
 }  // namespace viam
